feat: remove unnecessary file
